refactor: replace isPropertyTypeSupported
affects: @haulmont/jmix-react-core, @haulmont/jmix-react-ui

replace isPropertyTypeSupported() to jmix-react-core/util/metadata

ISSUES CLOSED: #58
